# postman/specs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SEAPEDIA API",
        "description": "Campus marketplace API — multi-role (buyer/seller/driver) with Sanctum Bearer auth.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local dev (Sail)"
        }
    ],
    "paths": {
...
        "/api/v1/buyer/wallet": {
            "get": {
                "tags": [
                    "Buyer Wallet"
                ],
                "summary": "Get the buyer's wallet balance and transaction history",
                "operationId": "c1b1f2c4275a714cbda88b6a68a2d0df",
                "responses": {
                    "200": {
                        "description": "Wallet balance and paginated transactions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/buyer/wallet/topup": {
            "post": {
                "tags": [
                    "Buyer Wallet"
                ],
                "summary": "Create a top-up (pending — poll the status endpoint to resolve it)",
                "operationId": "f9f55222d858a0ce5b2f1b5877e3e3c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Amount in IDR (integer, min 10000)",
                                        "type": "integer",
                                        "minimum": 10000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Topup created, pending"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/buyer/wallet/topup/{topup}": {
            "get": {
                "tags": [
                    "Buyer Wallet"
                ],
                "summary": "Poll a top-up — resolves to paid once the simulated gateway delay elapses",
                "operationId": "4b875dc3d70169be9747249625912b30",
                "responses": {
                    "200": {
                        "description": "Current top-up state"
                    },
                    "403": {
                        "description": "Not the owning buyer"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/catalog": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "List active products in the public catalog",
                "description": "Public catalog (TDD §8 \"GET /api/v1/catalog\") — same active-only data\nas the web catalog, both backed by `CatalogService`.",
                "operationId": "3f7c1e9e38a0efe2f0a51e39cad1a706",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Filter by product name (max 200 chars)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 200
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of active products"
                    }
                }
            }
        },
        "/api/v1/catalog/{slug}": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Get a single active product by slug",
                "description": "Public product detail by slug (TDD §8 \"GET /api/v1/catalog/{slug}\").",
                "operationId": "78e0afa9ea9c04d843e698462349ad43",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/v1/buyer/checkout/preview": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Preview the §5.2 money breakdown for a delivery method",
                "operationId": "35a98e74403eac3dadac0baa042e94ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "delivery_method"
                                ],
                                "properties": {
                                    "delivery_method": {
                                        "type": "string",
                                        "enum": [
                                            "pickup",
                                            "delivery"
                                        ]
                                    },
                                    "promo_code": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 32
                                    },
                                    "voucher_code": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 32
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "subtotal/discount/tax/delivery_fee/grand_total + balance check"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/buyer/checkout": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Commit checkout: charge wallet, reduce stock, create the order",
                "operationId": "c8a8b4b159192f35351c36f5bace341e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address_id",
                                    "delivery_method"
                                ],
                                "properties": {
                                    "address_id": {
                                        "type": "integer"
                                    },
                                    "delivery_method": {
                                        "type": "string",
                                        "enum": [
                                            "pickup",
                                            "delivery"
                                        ]
                                    },
                                    "promo_code": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 32
                                    },
                                    "voucher_code": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 32
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created in Sedang Dikemas"
                    },
                    "422": {
                        "description": "Insufficient stock or insufficient wallet balance"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum token — obtain from POST /api/v1/login",
                "scheme": "bearer"
            }
        }
    },
...
}